@@ -12,34 +12,35 @@
     "format": "prettier --write src/"
   },
   "dependencies": {
-    "@apidevtools/json-schema-ref-parser": "^10.1.0",
-    "@bytemd/plugin-gfm": "^1.21.0",
-    "@bytemd/plugin-highlight": "^1.21.0",
-    "@bytemd/vue-next": "^1.21.0",
-    "axios": "^1.5.0",
-    "form-data": "^4.0.0",
-    "monaco-editor": "^0.41.0",
-    "pinia": "^2.1.6",
-    "pinia-plugin-persistedstate": "^3.2.0",
-    "vue": "^3.3.4",
-    "vue-router": "^4.2.4"
+    "@apidevtools/json-schema-ref-parser": "11.6.3",
+    "@bytemd/plugin-gfm": "1.21.0",
+    "@bytemd/plugin-highlight": "1.21.0",
+    "@bytemd/vue-next": "1.21.0",
+    "axios": "1.7.2",
+    "bytemd": "^1.21.0",
+    "form-data": "4.0.0",
+    "monaco-editor": "0.49.0",
+    "pinia": "2.1.7",
+    "pinia-plugin-persistedstate": "^3.2.1",
+    "vue": "3.4.27",
+    "vue-router": "4.3.2"
   },
   "devDependencies": {
-    "@arco-design/web-vue": "^2.51.0",
-    "@rushstack/eslint-patch": "^1.3.2",
-    "@tsconfig/node18": "^18.2.0",
-    "@types/node": "^18.17.5",
-    "@vitejs/plugin-vue": "^4.3.1",
-    "@vue/eslint-config-prettier": "^8.0.0",
-    "@vue/eslint-config-typescript": "^11.0.3",
-    "@vue/tsconfig": "^0.4.0",
-    "eslint": "^8.46.0",
-    "eslint-plugin-vue": "^9.16.1",
-    "npm-run-all": "^4.1.5",
-    "openapi-typescript-codegen": "^0.25.0",
-    "prettier": "^3.0.0",
-    "typescript": "~5.1.6",
-    "vite": "^4.4.9",
-    "vue-tsc": "^1.8.8"
+    "@arco-design/web-vue": "2.55.2",
+    "@rushstack/eslint-patch": "1.10.3",
+    "@tsconfig/node18": "18.2.4",
+    "@types/node": "20.13.0",
+    "@vitejs/plugin-vue": "5.0.5",
+    "@vue/eslint-config-prettier": "9.0.0",
+    "@vue/eslint-config-typescript": "13.0.0",
+    "@vue/tsconfig": "0.5.1",
+    "eslint": "8.56.0",
+    "eslint-plugin-vue": "9.26.0",
+    "npm-run-all": "4.1.5",
+    "openapi-typescript-codegen": "0.29.0",
+    "prettier": "3.3.0",
+    "typescript": "5.4.5",
+    "vite": "5.2.12",
+    "vue-tsc": "2.0.19"
   }
 }
